requirements.txt
awscli==1.11.39
boto3==1.4.4
botocore==1.5.2
cent==2.0.2
colorama==0.3.7
dj-database-url==0.4.2
Django==1.10.5
django-classy-tags==0.8.0
django-cookie-law==1.0.13
django-datatables-view==1.14.0
django-filter==1.0.1
django-instant==0.2.1
django-registration-redux==1.4
django-smart-selects==1.2.7
django-storages==1.5.2
django-tables2==1.3.0
django-templateaddons3==1.0
django-tz-detect==0.2.8
djangorestframework==3.5.3
docutils==0.13.1
future==0.16.0
futures==3.0.5
httpie==0.9.9
isodate==0.5.4
jmespath==0.9.0
oauthlib==2.0.1
pdfkit==0.6.1
psycopg2==2.6.2
pyasn1==0.1.9
pycryptodome==3.4.3
pycryptodomex==3.4.3
Pygments==2.1.3
pyjwkest==1.3.2
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2016.10
PyYAML==3.12
requests==2.12.5
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
sqlparse==0.2.2
waitress==1.0.1
whitenoise==3.2.3
wkhtmltopdf-pack==0.12.3.0
